Clean: Relocate STEX RoundingHasher
