+++ src/dewarping/CMakeLists.txt
@@ -27,7 +27,6 @@ SET(
         STEX_PlugHoles.h
         STEX_PolygonRasterizer.cpp STEX_PolygonRasterizer.h
         STEX_RasterOpGeneric.h
-        STEX_RoundingHasher.cpp STEX_RoundingHasher.h
         STEX_ToPoint.h
         STEX_ToVec.h
         STEX_VecNT.h
