+++ src/foundation/RoundingHasher.h
